refactor(db): extract tasks/plans/drones/assets/teams — decomposition Wave 4
Wave 4: the 5 big entity modules. drones carries master's real renderJobForDrone
(SHELL_META + admin-gated setup, #154/#172) verbatim. Bodies verbatim; export* barrel.
Scope-trap re-imports added for resident callers. db.js: 2042 -> 950 (-1092). 28/31
modules. Gate: manifest --check green (308); suite 45 files / 310.

Executor: GLM-5.2 (agentic, gate-driven). Reviewed + verified by m5Max.
