+++ server/db.js
@@ -18,6 +18,16 @@ import { ensureDefaultChannels, getChannelsByUser, getUnreadCounts, isChannelMem
 import { listPluginRecords } from './db/plugins.js';
 import { getLatestSavepoint } from './db/savepoints.js';
 import { buildCalibrationBlock, createNodeProfile, getNodeProfile, seedPlatformProfiles, updateNodeProfile } from './db/node-profiles.js';
+// Wave 4 re-imports: resident boot/overview/workqueue/initDB functions still call
+// these now-extracted entity helpers by bare name. `export *` re-exports them to
+// consumers but does NOT bind them in THIS module's lexical scope (same trap the
+// earlier waves hit), so import them explicitly. Acyclic — each db/*.js imports
+// only ./db/core.js + sibling db/*.js, never this barrel (../db.js).
+import { listTasks, listTasksNeedingApproval } from './db/tasks.js';
+import { listPlans } from './db/plans.js';
+import { listAssets } from './db/assets.js';
+import { listDrones, listDroneJobs, seedDefaultJobTemplates } from './db/drones.js';
+import { getTeamsForUser } from './db/teams.js';
 
 export { getDB };
 
@@ -44,6 +54,11 @@ export * from './db/webhooks.js';
 export * from './db/channels.js';
 export * from './db/agent-profiles.js';
 export * from './db/node-profiles.js';
+export * from './db/tasks.js';
+export * from './db/plans.js';
+export * from './db/drones.js';
+export * from './db/assets.js';
+export * from './db/teams.js';
 
 export function initDB() {
   initDBConnection();
@@ -54,194 +69,6 @@ export function initDB() {
 }
 
 // =============== MYCELIUM PLATFORM ===============
-
-// -- Tasks --
-
-export function createTask(title, description, projectId, requester, priority, tags) {
-  var result = stmt('dvCreateTask', `INSERT INTO tasks (title, description, project_id, requester, priority, tags)
-    VALUES (?, ?, ?, ?, ?, ?) RETURNING id`).get(title, description || '', projectId || '', requester, priority || 'normal', tags || '[]');
-  return result.id;
-}
-
-export function getTask(id) {
-  return stmt('dvGetTask', 'SELECT * FROM tasks WHERE id = ?').get(id);
-}
-
-export function listTasks(filters) {
-  var where = ['1=1'];
-  var params = [];
-  if (filters.project_id) { where.push('project_id = ?'); params.push(filters.project_id); }
-  if (filters.status) { where.push('status = ?'); params.push(filters.status); }
-  if (filters.assignee) { where.push('assignee = ?'); params.push(filters.assignee); }
-  if (filters.requester) { where.push('requester = ?'); params.push(filters.requester); }
-  if (filters.priority) { where.push('priority = ?'); params.push(filters.priority); }
-  var limit = Math.min(filters.limit || 50, 500);
-  var offset = filters.offset || 0;
-  params.push(limit, offset);
-  return db.prepare('SELECT * FROM tasks WHERE ' + where.join(' AND ') + ' ORDER BY updated_at DESC LIMIT ? OFFSET ?').all(...params);
-}
-
-export function updateTask(id, fields) {
-  // Pre-process transformed fields
-  var f = Object.assign({}, fields);
-  if (f.assignee !== undefined && f.assignee !== null) f.assignee = resolveAssignee(f.assignee);
-  if (f.needs_approval !== undefined) f.needs_approval = f.needs_approval ? 1 : 0;
-  if (f.blocked_by !== undefined) f.blocked_by = JSON.stringify(f.blocked_by);
-  if (f.blocks !== undefined) f.blocks = JSON.stringify(f.blocks);
-  buildUpdate('tasks', id, f, ['title', 'description', 'status', 'assignee', 'priority', 'tags', 'needs_approval', 'blocked_by', 'blocks', 'branch', 'pr_url', 'repo', 'review_metadata'], { updatedAt: true });
-}
-
-// -- Task dependencies --
-
-export function setTaskDependency(taskId, blockedById) {
-  var task = getTask(taskId);
-  var blocker = getTask(blockedById);
-  if (!task || !blocker) return false;
-
-  var blockedBy = [];
-  try { blockedBy = JSON.parse(task.blocked_by || '[]'); } catch (e) { console.warn('[mycelium] JSON parse failed for task.blocked_by (task: ' + taskId + '):', e.message); }
-  if (blockedBy.indexOf(blockedById) === -1) {
-    blockedBy.push(blockedById);
-    db.prepare("UPDATE tasks SET blocked_by = ? WHERE id = ?").run(JSON.stringify(blockedBy), taskId);
-  }
-
-  var blocks = [];
-  try { blocks = JSON.parse(blocker.blocks || '[]'); } catch (e) { console.warn('[mycelium] JSON parse failed for task.blocks (task: ' + blockedById + '):', e.message); }
-  if (blocks.indexOf(taskId) === -1) {
-    blocks.push(taskId);
-    db.prepare("UPDATE tasks SET blocks = ? WHERE id = ?").run(JSON.stringify(blocks), blockedById);
-  }
-  return true;
-}
-
-export function resolveTaskDependencies(completedTaskId) {
-  var task = getTask(completedTaskId);
-  if (!task) return [];
-  var blocks = [];
-  try { blocks = JSON.parse(task.blocks || '[]'); } catch (e) { console.warn('[mycelium] JSON parse failed for task.blocks (task: ' + completedTaskId + '):', e.message); }
-
-  var unblocked = [];
-  for (var blockedId of blocks) {
-    var blocked = getTask(blockedId);
-    if (!blocked) continue;
-    var deps = [];
-    try { deps = JSON.parse(blocked.blocked_by || '[]'); } catch (e) { console.warn('[mycelium] JSON parse failed for task.blocked_by (task: ' + blockedId + '):', e.message); }
-    deps = deps.filter(function (d) { return d !== completedTaskId; });
-    db.prepare("UPDATE tasks SET blocked_by = ? WHERE id = ?").run(JSON.stringify(deps), blockedId);
-    if (deps.length === 0) unblocked.push(blockedId);
-  }
-  return unblocked;
-}
-
-// -- Task approval --
-
-export function approveTask(taskId, approvedBy) {
-  db.prepare("UPDATE tasks SET approved_by = ?, approved_at = datetime('now'), updated_at = datetime('now') WHERE id = ?").run(approvedBy, taskId);
-}
-
-export function listTasksNeedingApproval() {
-  return db.prepare("SELECT * FROM tasks WHERE needs_approval = 1 AND approved_by IS NULL AND status != 'done' ORDER BY updated_at DESC").all();
-}
-
-// -- Task Comments --
-
-export function addTaskComment(taskId, author, content) {
-  var result = db.prepare(
-    "INSERT INTO task_comments (task_id, author, content) VALUES (?, ?, ?) RETURNING *"
-  ).get(taskId, author, content);
-  return result;
-}
-
-export function getTaskComments(taskId) {
-  return db.prepare(
-    "SELECT * FROM task_comments WHERE task_id = ? ORDER BY created_at ASC"
-  ).all(taskId);
-}
-
-export function getTaskComment(commentId) {
-  return db.prepare("SELECT * FROM task_comments WHERE id = ?").get(commentId);
-}
-
-export function deleteTaskComment(commentId) {
-  var result = db.prepare("DELETE FROM task_comments WHERE id = ?").run(commentId);
-  return result.changes > 0;
-}
-
-// -- Task Deliverables --
-// The agent's final output, distinct from the task_comments status thread.
-// Append-only; getTaskDeliverables returns all attempts oldest-first.
-
-export function addTaskDeliverable(taskId, author, kind, format, content, flags) {
-  var result = db.prepare(
-    "INSERT INTO task_deliverables (task_id, author, kind, format, content, flags) VALUES (?, ?, ?, ?, ?, ?) RETURNING *"
-  ).get(taskId, author, kind || 'report', format || 'markdown', content, flags || '');
-  return result;
-}
-
-export function getTaskDeliverables(taskId) {
-  return db.prepare(
-    "SELECT * FROM task_deliverables WHERE task_id = ? ORDER BY created_at ASC, id ASC"
-  ).all(taskId);
-}
-
-export function deleteTask(id) {
-  db.prepare("DELETE FROM task_comments WHERE task_id = ?").run(id);
-  var result = db.prepare("DELETE FROM tasks WHERE id = ?").run(id);
-  return result.changes > 0;
-}
-
-// -- Plan Step Comments --
-
-export function addPlanStepComment(stepId, planId, author, content) {
-  var result = db.prepare(
-    "INSERT INTO plan_step_comments (step_id, plan_id, author, content) VALUES (?, ?, ?, ?) RETURNING *"
-  ).get(stepId, planId, author, content);
-  return result;
-}
-
-export function getPlanStepComments(stepId) {
-  return db.prepare(
-    "SELECT * FROM plan_step_comments WHERE step_id = ? ORDER BY created_at ASC"
-  ).all(stepId);
-}
-
-// -- Assets --
-
-export function createAsset(name, type, projectId, status, assetPath, metadata, requester) {
-  var result = stmt('dvCreateAsset', `INSERT INTO assets (name, type, project_id, status, path, metadata, requester)
-    VALUES (?, ?, ?, ?, ?, ?, ?) RETURNING id`).get(name, type || 'sprite', projectId || 'shared', status || 'requested', assetPath || '', metadata || '{}', requester || '');
-  return result.id;
-}
-
-export function getAsset(id) {
-  return stmt('dvGetAsset', 'SELECT * FROM assets WHERE id = ?').get(id);
-}
-
-export function listAssets(filters) {
-  var where = ['1=1'];
-  var params = [];
-  if (filters.project_id) { where.push('project_id = ?'); params.push(filters.project_id); }
-  if (filters.type) { where.push('type = ?'); params.push(filters.type); }
-  if (filters.status) { where.push('status = ?'); params.push(filters.status); }
-  var limit = Math.min(filters.limit || 50, 500);
-  var offset = filters.offset || 0;
-  params.push(limit, offset);
-  return db.prepare('SELECT * FROM assets WHERE ' + where.join(' AND ') + ' ORDER BY updated_at DESC LIMIT ? OFFSET ?').all(...params);
-}
-
-export function updateAsset(id, fields) {
-  // Returns {changes} for callers that check result.changes
-  var changed = buildUpdate('assets', id, fields, ['status', 'path', 'metadata', 'file_path', 'download_url', 'requested_by', 'assigned_to', 'drone_job_id', 'prompt'], { updatedAt: true });
-  return { changes: changed ? 1 : 0 };
-}
-
-export function deleteAsset(id) {
-  return db.prepare('DELETE FROM assets WHERE id = ?').run(id);
-}
-
-export function listAssetsByDroneJob(droneJobId) {
-  return db.prepare('SELECT * FROM assets WHERE drone_job_id = ?').all(droneJobId);
-}
 
 // -- Widgets --
 
@@ -934,702 +761,7 @@ export function getNextUnassignedPlanStep(teamProjectIds) {
   ).get(...params) || null;
 }
 
-// -- Auto-task from asset request --
-
-var _autoTaskFromAsset = null;
-
-export function initTransactions() {
-  _autoTaskFromAsset = db.transaction(function (assetId, projectId, requester) {
-    var agents = db.prepare("SELECT id FROM agents WHERE capabilities LIKE '%assets%'").all();
-    var assignee = agents.length > 0 ? agents[0].id : null;
-
-    var asset = getAsset(assetId);
-    if (!asset) return null;
-
-    var taskId = createTask(
-      'Generate asset: ' + asset.name,
-      'Auto-created from asset request #' + assetId + '. Type: ' + asset.type + '. Project: ' + projectId,
-      projectId,
-      requester,
-      'normal',
-      JSON.stringify(['auto', 'assets'])
-    );
-
-    db.prepare("UPDATE tasks SET assignee = ?, linked_asset_id = ? WHERE id = ?").run(assignee, assetId, taskId);
-
-    return { task_id: taskId, assignee: assignee };
-  });
-}
-
-export function autoTaskFromAsset(assetId, projectId, requester) {
-  if (!_autoTaskFromAsset) return null;
-  return _autoTaskFromAsset(assetId, projectId, requester);
-}
-
-// -- Plans --
-
-export function createPlan(title, description, projectId, owner, priority, tags, createdBy) {
-  var result = db.prepare(
-    "INSERT INTO plans (title, description, project_id, owner, priority, tags, created_by) VALUES (?, ?, ?, ?, ?, ?, ?) RETURNING id"
-  ).get(title, description || '', projectId || '', owner || '', priority || 'normal', tags || '[]', createdBy || '');
-  return result.id;
-}
-
-export function getPlan(id) {
-  var plan = db.prepare("SELECT * FROM plans WHERE id = ?").get(id);
-  if (!plan) return null;
-  var steps = db.prepare("SELECT * FROM plan_steps WHERE plan_id = ? ORDER BY step_order, id").all(id);
-  // Batch-fetch all comments for this plan and group by step
-  var allComments = db.prepare("SELECT * FROM plan_step_comments WHERE plan_id = ? ORDER BY created_at ASC").all(id);
-  var commentsByStep = {};
-  for (var c of allComments) {
-    if (!commentsByStep[c.step_id]) commentsByStep[c.step_id] = [];
-    commentsByStep[c.step_id].push(c);
-  }
-  for (var s of steps) {
-    s.comments = commentsByStep[s.id] || [];
-  }
-  var total = steps.length;
-  var completed = steps.filter(function (s) { return s.status === 'completed'; }).length;
-  plan.steps = steps;
-  plan.progress = { total: total, completed: completed, percent: total > 0 ? Math.round(completed / total * 100) : 0 };
-  return plan;
-}
-
-export function listPlans(filters) {
-  var where = ['1=1'];
-  var params = [];
-  if (filters.project_id) { where.push('project_id = ?'); params.push(filters.project_id); }
-  if (filters.status) { where.push('status = ?'); params.push(filters.status); }
-  if (filters.owner) { where.push('owner = ?'); params.push(filters.owner); }
-  if (filters.exclude_status) { where.push('status != ?'); params.push(filters.exclude_status); }
-  var limit = Math.min(filters.limit || 50, 500);
-  var offset = filters.offset || 0;
-  params.push(limit, offset);
-  var plans = db.prepare('SELECT * FROM plans WHERE ' + where.join(' AND ') + ' ORDER BY updated_at DESC LIMIT ? OFFSET ?').all(...params);
-  if (plans.length > 0) {
-    var planIds = plans.map(function (p) { return p.id; });
-    var placeholders = planIds.map(function () { return '?'; }).join(',');
-    var allSteps = db.prepare("SELECT plan_id, id, status, title, assignee, step_order FROM plan_steps WHERE plan_id IN (" + placeholders + ") ORDER BY step_order ASC").all(...planIds);
-    var stepsByPlan = {};
-    for (var s of allSteps) {
-      if (!stepsByPlan[s.plan_id]) stepsByPlan[s.plan_id] = [];
-      stepsByPlan[s.plan_id].push(s);
-    }
-    for (var p of plans) {
-      var steps = stepsByPlan[p.id] || [];
-      p.steps = steps;
-      var total = steps.length;
-      var completed = steps.filter(function (st) { return st.status === 'completed'; }).length;
-      p.step_count = total;
-      p.progress = { total: total, completed: completed, percent: total > 0 ? Math.round(completed / total * 100) : 0 };
-      var current = steps.find(function (st) { return st.status === 'in_progress'; }) ||
-                    steps.find(function (st) { return st.status === 'pending'; });
-      p.current_step = current ? current.title : null;
-    }
-  }
-  return plans;
-}
-
-export function updatePlan(id, fields) {
-  var f = Object.assign({}, fields);
-  if (f.tags !== undefined && typeof f.tags !== 'string') f.tags = JSON.stringify(f.tags);
-  buildUpdate('plans', id, f, ['title', 'description', 'status', 'owner', 'priority', 'tags', 'project_id'], { updatedAt: true });
-}
-
-export function deletePlan(id) {
-  db.prepare("DELETE FROM plan_steps WHERE plan_id = ?").run(id);
-  db.prepare("DELETE FROM plans WHERE id = ?").run(id);
-}
-
-export function createPlanStep(planId, title, description, assignee, phase) {
-  assignee = resolveAssignee(assignee);
-  var maxOrder = db.prepare("SELECT MAX(step_order) as m FROM plan_steps WHERE plan_id = ?").get(planId);
-  var order = (maxOrder && maxOrder.m !== null) ? maxOrder.m + 1 : 0;
-  var result = db.prepare(
-    "INSERT INTO plan_steps (plan_id, step_order, title, description, assignee, phase) VALUES (?, ?, ?, ?, ?, ?) RETURNING id"
-  ).get(planId, order, title, description || '', assignee || null, phase || '');
-  db.prepare("UPDATE plans SET updated_at = datetime('now') WHERE id = ?").run(planId);
-  return result.id;
-}
-
-export function updatePlanStep(stepId, fields) {
-  if (fields.assignee !== undefined && fields.assignee !== null) fields.assignee = resolveAssignee(fields.assignee);
-  var extra = fields.status === 'completed' ? ["completed_at = datetime('now')"] : [];
-  buildUpdate('plan_steps', stepId, fields, ['title', 'description', 'status', 'assignee', 'linked_task_id', 'linked_branch', 'linked_pr_url', 'phase', 'step_order'], { updatedAt: true, extraSets: extra });
-  // Update parent plan's updated_at
-  var step = db.prepare("SELECT plan_id FROM plan_steps WHERE id = ?").get(stepId);
-  if (step) db.prepare("UPDATE plans SET updated_at = datetime('now') WHERE id = ?").run(step.plan_id);
-}
-
-// Bounded self-heal for a FAILED plan step. If the step still has retry budget,
-// reopen it + the phase it guards (the immediately-prior step_order) to 'pending'
-// and attach the failure `critique` to those reopened steps, so the work re-cycles
-// with the verifier's reasons fed forward. Returns {action:'retried'|'exhausted'|'none'}.
-// Runtime-agnostic: it only moves platform rows; the reopened steps flow through the
-// normal work queue to whatever agent/runner is assigned. The caller decides what to
-// do on 'exhausted' (block + escalate). maxAttempts is the caller's policy knob.
-export function autoRetryOrEscalatePlanStep(planId, stepId, maxAttempts, critique) {
-  var failed = db.prepare("SELECT id, step_order, attempt_count FROM plan_steps WHERE id = ? AND plan_id = ?").get(stepId, planId);
-  if (!failed) return { action: 'none' };
-  var attempts = failed.attempt_count || 0;
-  if (attempts >= maxAttempts) return { action: 'exhausted', attempts: attempts };
-  var nextAttempt = attempts + 1;
-  var steps = db.prepare("SELECT id, step_order, status FROM plan_steps WHERE plan_id = ?").all(planId);
-  // The phase this step guards = the greatest step_order below it (re-run the impl
-  // a verify step checks). Reopen those completed steps + the failed step itself.
-  var priorOrders = steps.map(function (s) { return s.step_order; }).filter(function (o) { return o < failed.step_order; });
-  var priorOrder = priorOrders.length ? Math.max.apply(null, priorOrders) : null;
-  var toReopen = steps.filter(function (s) {
-    return s.id === stepId || (priorOrder !== null && s.step_order === priorOrder && s.status === 'completed');
-  });
-  var tx = db.transaction(function () {
-    for (var s of toReopen) {
-      db.prepare("UPDATE plan_steps SET status = 'pending', completed_at = NULL, updated_at = datetime('now') WHERE id = ?").run(s.id);
-      if (critique && s.id !== stepId) {
-        addPlanStepComment(s.id, planId, '__system__',
-          '[auto-retry ' + nextAttempt + '/' + maxAttempts + '] Prior verification FAILED — fix before re-submitting:\n' + critique);
-      }
-    }
-    db.prepare("UPDATE plan_steps SET attempt_count = ?, updated_at = datetime('now') WHERE id = ?").run(nextAttempt, stepId);
-    db.prepare("UPDATE plans SET updated_at = datetime('now') WHERE id = ?").run(planId);
-  });
-  tx();
-  return { action: 'retried', attempt: nextAttempt, max: maxAttempts, reopened: toReopen.length };
-}
-
-export function deletePlanStep(stepId) {
-  var step = db.prepare("SELECT plan_id FROM plan_steps WHERE id = ?").get(stepId);
-  db.prepare("DELETE FROM plan_steps WHERE id = ?").run(stepId);
-  if (step) db.prepare("UPDATE plans SET updated_at = datetime('now') WHERE id = ?").run(step.plan_id);
-}
-
-export function reorderPlanSteps(planId, stepIds) {
-  var reorder = db.transaction(function () {
-    for (var i = 0; i < stepIds.length; i++) {
-      db.prepare("UPDATE plan_steps SET step_order = ? WHERE id = ? AND plan_id = ?").run(i, stepIds[i], planId);
-    }
-    db.prepare("UPDATE plans SET updated_at = datetime('now') WHERE id = ?").run(planId);
-  });
-  reorder();
-}
-
-export function completeLinkedPlanSteps(taskId) {
-  var steps = db.prepare("SELECT id, plan_id FROM plan_steps WHERE linked_task_id = ? AND status != 'completed'").all(taskId);
-  var affectedPlanIds = [];
-  for (var step of steps) {
-    db.prepare("UPDATE plan_steps SET status = 'completed', completed_at = datetime('now'), updated_at = datetime('now') WHERE id = ?").run(step.id);
-    if (affectedPlanIds.indexOf(step.plan_id) === -1) affectedPlanIds.push(step.plan_id);
-  }
-  // Check if any affected plans are now fully complete
-  var completedPlans = [];
-  for (var planId of affectedPlanIds) {
-    var remaining = db.prepare("SELECT COUNT(*) as c FROM plan_steps WHERE plan_id = ? AND status NOT IN ('completed', 'skipped')").get(planId);
-    if (remaining.c === 0) {
-      db.prepare("UPDATE plans SET status = 'completed', updated_at = datetime('now') WHERE id = ? AND status = 'active'").run(planId);
-      completedPlans.push(planId);
-    } else {
-      db.prepare("UPDATE plans SET updated_at = datetime('now') WHERE id = ?").run(planId);
-    }
-  }
-  return { steps_completed: steps.length, plans_completed: completedPlans };
-}
-
 // -- Studio Users --
-
-// -- Drone Jobs --
-
-export function createDroneJob(title, command, inputData, requires, requester, priority, workspaceRepo, workspaceBranch, profileId) {
-  var result = db.prepare(
-    "INSERT INTO drone_jobs (title, command, input_data, requires, requester, priority, workspace_repo, workspace_branch, profile_id) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?) RETURNING id"
-  ).get(
-    title,
-    command || '',
-    typeof inputData === 'string' ? inputData : JSON.stringify(inputData || {}),
-    typeof requires === 'string' ? requires : JSON.stringify(requires || ['cpu']),
-    requester,
-    priority || 0,
-    workspaceRepo || null,
-    workspaceBranch || 'main',
-    profileId || null
-  );
-  return result.id;
-}
-
-export function getDroneJob(id) {
-  return db.prepare("SELECT * FROM drone_jobs WHERE id = ?").get(id);
-}
-
-export function claimDroneJob(droneId, capabilities) {
-  var caps = Array.isArray(capabilities) ? capabilities : [];
-
-  return db.transaction(function () {
-    var pending = db.prepare(
-      "SELECT * FROM drone_jobs WHERE status = 'pending' ORDER BY priority DESC, created_at ASC"
-    ).all();
-
-    for (var i = 0; i < pending.length; i++) {
-      var job = pending[i];
-      var reqs = [];
-      try { reqs = JSON.parse(job.requires || '["cpu"]'); } catch (e) { console.warn('[mycelium] JSON parse failed for job.requires (job: ' + job.id + '):', e.message); reqs = ['cpu']; }
-      var matched = reqs.every(function (r) { return caps.indexOf(r) !== -1; });
-      if (!matched) continue;
-
-      if (job.profile_id) {
-        var assignment = db.prepare(
-          "SELECT setup_done FROM drone_profile_assignments WHERE drone_id = ? AND profile_id = ?"
-        ).get(droneId, job.profile_id);
-        if (!assignment || !assignment.setup_done) continue;
-      }
-
-      var result = db.prepare(
-        "UPDATE drone_jobs SET status = 'claimed', drone_id = ?, started_at = datetime('now') WHERE id = ? AND status = 'pending'"
-      ).run(droneId, job.id);
-      if (result.changes > 0) return getDroneJob(job.id);
-    }
-    return null;
-  })();
-}
-
-export function updateDroneJob(id, fields) {
-  var f = Object.assign({}, fields);
-  if (f.input_data !== undefined && typeof f.input_data !== 'string') f.input_data = JSON.stringify(f.input_data);
-  if (f.result_data !== undefined && typeof f.result_data !== 'string') f.result_data = JSON.stringify(f.result_data);
-  buildUpdate('drone_jobs', id, f, ['status', 'command', 'input_data', 'result_url', 'result_data', 'error', 'completed_at', 'drone_id', 'started_at']);
-}
-
-export function listDroneJobs(filters) {
-  var where = ['1=1'];
-  var params = [];
-  if (filters.status) { where.push('status = ?'); params.push(filters.status); }
-  if (filters.drone_id) { where.push('drone_id = ?'); params.push(filters.drone_id); }
-  if (filters.requester) { where.push('requester = ?'); params.push(filters.requester); }
-  var limit = Math.min(filters.limit || 50, 200);
-  var offset = filters.offset || 0;
-  params.push(limit, offset);
-  return db.prepare('SELECT * FROM drone_jobs WHERE ' + where.join(' AND ') + ' ORDER BY created_at DESC LIMIT ? OFFSET ?').all(...params);
-}
-
-// Bug #137: Release stale claimed jobs (claimed >1 hour ago, not completed)
-// If droneId provided, scoped to that drone. Otherwise releases ALL stale jobs.
-export function releaseStaleClaimedJobs(droneId) {
-  var staleJobs;
-  if (droneId) {
-    staleJobs = db.prepare(
-      "SELECT * FROM drone_jobs WHERE status = 'claimed' AND drone_id = ? AND started_at < datetime('now', '-1 hour')"
-    ).all(droneId);
-  } else {
-    staleJobs = db.prepare(
-      "SELECT * FROM drone_jobs WHERE status = 'claimed' AND started_at < datetime('now', '-1 hour')"
-    ).all();
-  }
-  var msg = droneId
-    ? '[stale_timeout] Job was claimed for >1 hour with no completion. Auto-failed on drone restart.'
-    : '[stale_timeout] Job was claimed for >1 hour with no progress. Auto-failed by server.';
-  for (var job of staleJobs) {
-    db.prepare(
-      "UPDATE drone_jobs SET status = 'failed', error = ?, completed_at = datetime('now') WHERE id = ?"
-    ).run(msg, job.id);
-  }
-  return staleJobs;
-}
-
-export function listDrones() {
-  return db.prepare("SELECT id, name, project_id, status, working_on, last_heartbeat, capabilities, created_at FROM agents WHERE project_id = 'drone' ORDER BY created_at").all();
-}
-
-export function pauseDrone(droneId) {
-  db.prepare(`UPDATE agents SET status = 'paused', working_on = 'Paused (GPU released)' WHERE id = ?`).run(droneId);
-  return { ok: true, status: 'paused' };
-}
-
-export function resumeDrone(droneId) {
-  db.prepare(`UPDATE agents SET status = 'online', working_on = '' WHERE id = ?`).run(droneId);
-  return { ok: true, status: 'online' };
-}
-
-export function getDroneStatus(droneId) {
-  var agent = db.prepare(`SELECT id, name, status, working_on, capabilities, last_heartbeat FROM agents WHERE id = ?`).get(droneId);
-  if (!agent) return null;
-  var pendingJobs = db.prepare(`SELECT COUNT(*) as count FROM drone_jobs WHERE status = 'pending'`).get();
-  return {
-    ...agent,
-    capabilities: JSON.parse(agent.capabilities || '[]'),
-    queued_jobs: pendingJobs.count
-  };
-}
-
-// -- Drone Profiles --
-
-export function createDroneProfile(id, name, description, requires, artifacts, setupScript, workspace, env) {
-  db.prepare(
-    "INSERT INTO drone_profiles (id, name, description, requires, artifacts, setup_script, workspace, env) VALUES (?, ?, ?, ?, ?, ?, ?, ?)"
-  ).run(
-    id, name, description || '',
-    typeof requires === 'string' ? requires : JSON.stringify(requires || {}),
-    typeof artifacts === 'string' ? artifacts : JSON.stringify(artifacts || []),
-    setupScript || '',
-    workspace || '',
-    typeof env === 'string' ? env : JSON.stringify(env || {})
-  );
-  return getDroneProfile(id);
-}
-
-export function getDroneProfile(id) {
-  return db.prepare("SELECT * FROM drone_profiles WHERE id = ?").get(id);
-}
-
-export function listDroneProfiles() {
-  return db.prepare("SELECT * FROM drone_profiles ORDER BY created_at").all();
-}
-
-export function updateDroneProfile(id, fields) {
-  var sets = [];
-  var values = [];
-  for (var key of ['name', 'description', 'setup_script', 'workspace']) {
-    if (fields[key] !== undefined) { sets.push(key + ' = ?'); values.push(fields[key]); }
-  }
-  for (var jsonKey of ['requires', 'artifacts', 'env']) {
-    if (fields[jsonKey] !== undefined) {
-      sets.push(jsonKey + ' = ?');
-      values.push(typeof fields[jsonKey] === 'string' ? fields[jsonKey] : JSON.stringify(fields[jsonKey]));
-    }
-  }
-  if (sets.length === 0) return getDroneProfile(id);
-  sets.push("updated_at = datetime('now')");
-  values.push(id);
-  db.prepare('UPDATE drone_profiles SET ' + sets.join(', ') + ' WHERE id = ?').run(...values);
-  // Invalidate setup_done for all drones assigned to this profile
-  db.prepare("UPDATE drone_profile_assignments SET setup_done = 0, checksum = '' WHERE profile_id = ?").run(id);
-  return getDroneProfile(id);
-}
-
-export function deleteDroneProfile(id) {
-  return db.prepare("DELETE FROM drone_profiles WHERE id = ?").run(id);
-}
-
-export function assignDroneProfile(droneId, profileId) {
-  db.prepare(
-    "INSERT OR REPLACE INTO drone_profile_assignments (drone_id, profile_id, setup_done, checksum) VALUES (?, ?, 0, '')"
-  ).run(droneId, profileId);
-}
-
-export function unassignDroneProfile(droneId, profileId) {
-  return db.prepare("DELETE FROM drone_profile_assignments WHERE drone_id = ? AND profile_id = ?").run(droneId, profileId);
-}
-
-export function getDroneProfileAssignments(droneId) {
-  return db.prepare(
-    "SELECT a.*, p.name, p.description, p.requires, p.artifacts, p.setup_script, p.workspace, p.env, p.updated_at as profile_updated_at " +
-    "FROM drone_profile_assignments a JOIN drone_profiles p ON a.profile_id = p.id WHERE a.drone_id = ? ORDER BY p.created_at"
-  ).all(droneId);
-}
-
-export function markProfileSetupDone(droneId, profileId, checksum) {
-  db.prepare(
-    "UPDATE drone_profile_assignments SET setup_done = 1, setup_at = datetime('now'), checksum = ? WHERE drone_id = ? AND profile_id = ?"
-  ).run(checksum || '', droneId, profileId);
-}
-
-export function getDronesWithProfile(profileId) {
-  return db.prepare(
-    "SELECT a.drone_id, a.setup_done, a.setup_at, a.checksum, ag.status, ag.last_heartbeat " +
-    "FROM drone_profile_assignments a JOIN agents ag ON a.drone_id = ag.id WHERE a.profile_id = ?"
-  ).all(profileId);
-}
-
-export function bulkCancelDroneJobs(statuses, olderThanDays) {
-  var placeholders = statuses.map(function () { return '?'; }).join(',');
-  var params = statuses.slice();
-  var where = 'status IN (' + placeholders + ')';
-  if (parseInt(olderThanDays) > 0) {
-    where += " AND completed_at < datetime('now', '-' || ? || ' days')";
-    params.push(String(parseInt(olderThanDays)));
-  }
-  var jobs = db.prepare('SELECT id, title, status FROM drone_jobs WHERE ' + where).all.apply(
-    db.prepare('SELECT id, title, status FROM drone_jobs WHERE ' + where), params
-  );
-  if (jobs.length > 0) {
-    var idPlaceholders = jobs.map(function () { return '?'; }).join(',');
-    var ids = jobs.map(function (j) { return j.id; });
-    db.prepare("UPDATE drone_jobs SET status = 'cancelled' WHERE id IN (" + idPlaceholders + ')').run.apply(
-      db.prepare("UPDATE drone_jobs SET status = 'cancelled' WHERE id IN (" + idPlaceholders + ')'), ids
-    );
-  }
-  return jobs;
-}
-
-// -- Job Templates --
-
-// Seed the 3d_print template so new instances support printer drones out of the box.
-export function seedDefaultJobTemplates() {
-  var existing = db.prepare("SELECT id FROM job_templates WHERE id = '3d_print'").get();
-  if (!existing) {
-    db.prepare(
-      "INSERT INTO job_templates (id, name, project_id, requires, min_vram_gb, min_disk_gb) VALUES (?, ?, ?, ?, ?, ?)"
-    ).run('3d_print', '3D Print Job', '', '["3d_printer"]', 0, 1);
-    console.log('Seeded 3d_print job template');
-  }
-}
-
-export function createJobTemplate(id, fields) {
-  db.prepare(
-    "INSERT INTO job_templates (id, name, project_id, requires, min_vram_gb, min_disk_gb, python_deps, python_deps_install, artifacts, setup_repo, command_template, workspace_name) " +
-    "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)"
-  ).run(
-    id,
-    fields.name || id,
-    fields.project_id || '',
-    typeof fields.requires === 'string' ? fields.requires : JSON.stringify(fields.requires || ['cpu']),
-    fields.min_vram_gb || 0,
-    fields.min_disk_gb || 5,
-    typeof fields.python_deps === 'string' ? fields.python_deps : JSON.stringify(fields.python_deps || []),
-    fields.python_deps_install || '',
-    typeof fields.artifacts === 'string' ? fields.artifacts : JSON.stringify(fields.artifacts || []),
-    fields.setup_repo || '',
-    fields.command_template || '',
-    fields.workspace_name || ''
-  );
-  return getJobTemplate(id);
-}
-
-export function getJobTemplate(id) {
-  return db.prepare("SELECT * FROM job_templates WHERE id = ?").get(id);
-}
-
-export function listJobTemplates() {
-  return db.prepare("SELECT * FROM job_templates ORDER BY created_at").all();
-}
-
-export function updateJobTemplate(id, fields) {
-  var sets = [];
-  var values = [];
-  for (var key of ['name', 'project_id', 'python_deps_install', 'setup_repo', 'command_template', 'workspace_name']) {
-    if (fields[key] !== undefined) { sets.push(key + ' = ?'); values.push(fields[key]); }
-  }
-  for (var numKey of ['min_vram_gb', 'min_disk_gb']) {
-    if (fields[numKey] !== undefined) { sets.push(numKey + ' = ?'); values.push(fields[numKey]); }
-  }
-  for (var jsonKey of ['requires', 'python_deps', 'artifacts']) {
-    if (fields[jsonKey] !== undefined) {
-      sets.push(jsonKey + ' = ?');
-      values.push(typeof fields[jsonKey] === 'string' ? fields[jsonKey] : JSON.stringify(fields[jsonKey]));
-    }
-  }
-  if (sets.length === 0) return getJobTemplate(id);
-  values.push(id);
-  db.prepare('UPDATE job_templates SET ' + sets.join(', ') + ' WHERE id = ?').run(...values);
-  return getJobTemplate(id);
-}
-
-export function deleteJobTemplate(id) {
-  return db.prepare("DELETE FROM job_templates WHERE id = ?").run(id);
-}
-
-// -- Drone Diagnostics --
-
-export function updateDroneDiagnostics(agentId, diagnostics) {
-  var json = typeof diagnostics === 'string' ? diagnostics : JSON.stringify(diagnostics);
-  db.prepare("UPDATE agents SET system_diagnostics = ? WHERE id = ?").run(json, agentId);
-}
-
-export function getDroneDiagnostics(agentId) {
-  var row = db.prepare("SELECT system_diagnostics FROM agents WHERE id = ?").get(agentId);
-  if (!row) return null;
-  try { return JSON.parse(row.system_diagnostics || '{}'); } catch (e) { return {}; }
-}
-
-// -- Platform Resolver + Command Renderer --
-
-export function renderJobForDrone(templateId, droneId, inputData) {
-  var template = getJobTemplate(templateId);
-  if (!template) return { error: 'Template not found: ' + templateId };
-
-  var diag = getDroneDiagnostics(droneId);
-  if (!diag || Object.keys(diag).length === 0) {
-    // Fall back to savepoint system_info
-    var savepoint = getLatestSavepoint(droneId);
-    if (savepoint) {
-      try {
-        var snapshot = JSON.parse(savepoint.state_snapshot || '{}');
-        diag = snapshot.system_info || {};
-      } catch (e) { diag = {}; }
-    }
-  }
-  if (!diag || Object.keys(diag).length === 0) {
-    return { error: 'No diagnostics available for drone ' + droneId + '. Drone must heartbeat first.' };
-  }
-
-  // Compatibility checks
-  var templateReqs = [];
-  try { templateReqs = JSON.parse(template.requires || '["cpu"]'); } catch (e) { templateReqs = ['cpu']; }
-
-  // Check GPU requirement
-  if (templateReqs.indexOf('gpu') !== -1) {
-    if (!diag.cuda_available && !diag.gpu_name) {
-      return { error: 'Template requires GPU but drone has none', incompatible: true };
-    }
-    if (template.min_vram_gb > 0 && diag.gpu_vram_gb && diag.gpu_vram_gb < template.min_vram_gb) {
-      return { error: 'Template requires ' + template.min_vram_gb + ' GB VRAM but drone has ' + diag.gpu_vram_gb + ' GB', incompatible: true };
-    }
-  }
-
-  // Check disk
-  if (template.min_disk_gb > 0 && diag.disk_free_gb && diag.disk_free_gb < template.min_disk_gb) {
-    return { error: 'Template requires ' + template.min_disk_gb + ' GB free disk but drone has ' + diag.disk_free_gb + ' GB', incompatible: true };
-  }
-
-  // Resolve platform vars
-  var isWindows = (diag.os || '').toLowerCase() === 'windows';
-  var pythonPath = diag.python_path || (isWindows ? 'python' : 'python3');
-  var home = diag.home || (isWindows ? 'C:/Users/' + (diag.username || 'user') : '/home/' + (diag.username || 'user'));
-  var workspaceName = template.workspace_name || templateId;
-  var workspace = home + '/.mycelium/workspaces/' + workspaceName;
-  var nullDev = isWindows ? 'NUL' : '/dev/null';
-  var pipInstall = pythonPath + ' -m pip install';
-  var pathSep = isWindows ? ';' : ':';
-
-  // Build setup steps
-  var setupSteps = [];
-
-  // Step 1: Download artifacts (handled by worker from input_data.artifacts)
-  var templateArtifacts = [];
-  try { templateArtifacts = JSON.parse(template.artifacts || '[]'); } catch (e) { templateArtifacts = []; }
-
-  // Step 2: Clone setup_repo if specified
-  if (template.setup_repo) {
-    var repoDir = workspace + '/ai-toolkit';
-    var checkClone = isWindows
-      ? 'if exist "' + repoDir.replace(/\//g, '\\') + '" (exit /b 0) else (exit /b 1)'
-      : 'test -d "' + repoDir + '"';
-    setupSteps.push({
-      name: 'Clone setup repo',
-      check: checkClone,
-      run: 'git clone "' + template.setup_repo + '" "' + repoDir + '"',
-      skip_if_check_passes: true,
-    });
-  }
-
-  // Step 3: Install CUDA torch (if GPU required)
-  if (templateReqs.indexOf('gpu') !== -1) {
-    setupSteps.push({
-      name: 'Check/install CUDA PyTorch',
-      check: pythonPath + ' -c "import torch; assert torch.cuda.is_available(), \'no cuda\'"',
-      run: pipInstall + ' torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124',
-      skip_if_check_passes: true,
-    });
-  }
-
-  // Step 4: Install python deps
-  if (template.python_deps_install) {
-    var depsList = [];
-    try { depsList = JSON.parse(template.python_deps || '[]'); } catch (e) { depsList = []; }
-    var importChecks = depsList.map(function (d) {
-      // PIL -> Pillow, map common package names
-      var mod = d === 'PIL' ? 'PIL' : d;
-      return 'import ' + mod;
-    }).join('; ');
-    setupSteps.push({
-      name: 'Install Python dependencies',
-      check: pythonPath + ' -c "' + importChecks + '"',
-      run: pipInstall + ' ' + template.python_deps_install,
-      skip_if_check_passes: true,
-    });
-  }
-
-  // Render command template
-  var vars = {
-    python: pythonPath,
-    pip_install: pipInstall,
-    workspace: workspace,
-    null_dev: nullDev,
-    path_sep: pathSep,
-  };
-  // Merge inputData vars. These are user-supplied and get interpolated into a
-  // command executed on the drone, so reject shell metacharacters (C-2): $ and
-  // backtick do command substitution even inside quotes; ; | & < > chain and
-  // redirect. Blocking them stops injection cross-OS without touching normal
-  // prompt text (letters, spaces, commas, apostrophes, parens all pass).
-  var SHELL_META = /[$`;|&<>\n\r\0]/;
-  if (inputData && typeof inputData === 'object') {
-    for (var k of Object.keys(inputData)) {
-      if (k.startsWith('_')) continue;
-      var v = inputData[k];
-      if (typeof v === 'string' && SHELL_META.test(v)) {
-        return { error: 'Input value for "' + k + '" contains disallowed shell characters ($ ` ; | & < > newline)' };
-      }
-      vars[k] = v;
-    }
-  }
-  var command = template.command_template;
-  for (var [varName, varVal] of Object.entries(vars)) {
-    command = command.replace(new RegExp('\\{\\{' + varName + '\\}\\}', 'g'), String(varVal));
-  }
-
-  return {
-    command: command,
-    setup_steps: setupSteps,
-    artifacts: templateArtifacts,
-    requires: templateReqs,
-    workspace: workspace,
-    workspace_name: workspaceName,
-    template_id: templateId,
-    drone_diagnostics: {
-      os: diag.os,
-      python_path: pythonPath,
-      gpu_name: diag.gpu_name,
-      gpu_vram_gb: diag.gpu_vram_gb,
-    },
-  };
-}
-
-export function checkDroneCompatibility(droneId) {
-  var templates = listJobTemplates();
-  var diag = getDroneDiagnostics(droneId);
-  if (!diag || Object.keys(diag).length === 0) {
-    var savepoint = getLatestSavepoint(droneId);
-    if (savepoint) {
-      try {
-        var snapshot = JSON.parse(savepoint.state_snapshot || '{}');
-        diag = snapshot.system_info || {};
-      } catch (e) { diag = {}; }
-    }
-  }
-  if (!diag || Object.keys(diag).length === 0) {
-    return { drone_id: droneId, error: 'No diagnostics available', compatible: [], incompatible: [] };
-  }
-
-  var compatible = [];
-  var incompatible = [];
-
-  for (var t of templates) {
-    var reqs = [];
-    try { reqs = JSON.parse(t.requires || '["cpu"]'); } catch (e) { reqs = ['cpu']; }
-    var issues = [];
-
-    if (reqs.indexOf('gpu') !== -1) {
-      if (!diag.cuda_available && !diag.gpu_name) {
-        issues.push('Requires GPU, none detected');
-      } else if (t.min_vram_gb > 0 && diag.gpu_vram_gb && diag.gpu_vram_gb < t.min_vram_gb) {
-        issues.push('Requires ' + t.min_vram_gb + ' GB VRAM, has ' + diag.gpu_vram_gb + ' GB');
-      }
-    }
-    if (t.min_disk_gb > 0 && diag.disk_free_gb && diag.disk_free_gb < t.min_disk_gb) {
-      issues.push('Requires ' + t.min_disk_gb + ' GB disk, has ' + diag.disk_free_gb + ' GB');
-    }
-
-    if (issues.length === 0) {
-      var notes = [];
-      if (diag.gpu_name) notes.push(diag.gpu_name + ' ' + (diag.gpu_vram_gb || '?') + ' GB VRAM');
-      if (diag.disk_free_gb) notes.push(diag.disk_free_gb + ' GB free disk');
-      compatible.push({ template: t.id, name: t.name, status: 'ready', notes: notes.join(', ') });
-    } else {
-      incompatible.push({ template: t.id, name: t.name, status: 'incompatible', reasons: issues });
-    }
-  }
-
-  return { drone_id: droneId, compatible: compatible, incompatible: incompatible };
-}
 
 // -- Init (no default seed data — new instances start blank) --
 
@@ -1816,227 +948,3 @@ export function getSlimOverview() {
 
   return { agents: agents, counts: counts, attention: attention, recent_activity: recent_activity };
 }
-
-// =============== TEAM SETTINGS ===============
-
-export function listTeamSettings(section) {
-  if (section) {
-    return db.prepare('SELECT * FROM team_settings WHERE section = ? ORDER BY key').all(section);
-  }
-  return db.prepare('SELECT * FROM team_settings ORDER BY section, key').all();
-}
-
-export function getTeamSetting(section, key) {
-  return db.prepare('SELECT * FROM team_settings WHERE section = ? AND key = ?').get(section, key);
-}
-
-export function upsertTeamSetting(section, key, value, updatedBy) {
-  var now = new Date().toISOString();
-  var valueStr = typeof value === 'object' ? JSON.stringify(value) : String(value);
-  db.prepare(
-    "INSERT INTO team_settings (section, key, value, updated_at, updated_by) VALUES (?, ?, ?, ?, ?) " +
-    "ON CONFLICT(section, key) DO UPDATE SET value = excluded.value, updated_at = excluded.updated_at, updated_by = excluded.updated_by"
-  ).run(section, key, valueStr, now, updatedBy || '');
-  syncTeamSettingsToProfile();
-  return getTeamSetting(section, key);
-}
-
-export function deleteTeamSetting(section, key) {
-  var result = db.prepare('DELETE FROM team_settings WHERE section = ? AND key = ?').run(section, key);
-  syncTeamSettingsToProfile();
-  return result;
-}
-
-export function getAllTeamSettingsGrouped() {
-  var rows = listTeamSettings();
-  var grouped = {};
-  for (var row of rows) {
-    if (!grouped[row.section]) grouped[row.section] = {};
-    try {
-      grouped[row.section][row.key] = JSON.parse(row.value);
-    } catch (e) {
-      grouped[row.section][row.key] = row.value;
-    }
-  }
-  return grouped;
-}
-
-export function syncTeamSettingsToProfile() {
-  var settings = getAllTeamSettingsGrouped();
-  var profileId = 'customer-agent';
-  var existing = getNodeProfile(profileId);
-
-  var updates = {};
-
-  // Guardrails → direct profile mapping
-  var guardrails = settings.guardrails || {};
-  if (guardrails.tool_whitelist) updates.tool_whitelist = guardrails.tool_whitelist;
-  if (guardrails.repo_list) updates.repo_list = guardrails.repo_list;
-  if (guardrails.md_checkpoints) updates.md_checkpoints = guardrails.md_checkpoints;
-  if (guardrails.md_blocklist) updates.md_blocklist = guardrails.md_blocklist;
-
-  // Build rules from multiple sections
-  var rules = {};
-  if (existing) {
-    try { rules = typeof existing.rules === 'object' ? existing.rules : JSON.parse(existing.rules || '{}'); } catch (e) { rules = {}; }
-  }
-
-  // Coding standards → rule
-  var coding = settings.coding_standards || {};
-  if (Object.keys(coding).length > 0) {
-    var parts = [];
-    if (coding.languages && coding.languages.length) parts.push('Languages: ' + coding.languages.join(', '));
-    if (coding.linter) parts.push('Linter: ' + coding.linter);
-    if (coding.formatter) parts.push('Formatter: ' + coding.formatter);
-    if (coding.test_framework) parts.push('Tests: ' + coding.test_framework);
-    if (coding.style_notes) parts.push(coding.style_notes);
-    rules.coding_standards = { severity: 'high', description: parts.join('. ') };
-
-    // Also add language names to md_checkpoints
-    if (coding.languages && coding.languages.length) {
-      var checkpoints = updates.md_checkpoints || (existing && existing.md_checkpoints) || [];
-      if (typeof checkpoints === 'string') try { checkpoints = JSON.parse(checkpoints); } catch (e) { checkpoints = []; }
-      for (var lang of coding.languages) {
-        if (checkpoints.indexOf(lang) === -1) checkpoints.push(lang);
-      }
-      updates.md_checkpoints = checkpoints;
-    }
-  }
-
-  // Deploy workflow → rule
-  var deploy = settings.deploy_workflow || {};
-  if (Object.keys(deploy).length > 0) {
-    var deployParts = [];
-    if (deploy.stages && deploy.stages.length) deployParts.push('Stages: ' + deploy.stages.join(' \u2192 '));
-    if (deploy.deploy_method) deployParts.push('Method: ' + deploy.deploy_method);
-    if (deploy.pr_requirements) deployParts.push('PR: ' + JSON.stringify(deploy.pr_requirements));
-    rules.deploy_workflow = { severity: 'high', description: deployParts.join('. ') };
-  }
-
-  // Team rules → rule
-  var teamRules = settings.team_rules || {};
-  if (Object.keys(teamRules).length > 0) {
-    var trParts = [];
-    if (teamRules.communication_style) trParts.push('Style: ' + teamRules.communication_style);
-    if (teamRules.timezone) trParts.push('TZ: ' + teamRules.timezone);
-    if (teamRules.working_hours) trParts.push('Hours: ' + teamRules.working_hours);
-    rules.team_rules = { severity: 'medium', description: trParts.join('. ') };
-  }
-
-  // Custom guardrail rules
-  if (guardrails.custom_rules && Array.isArray(guardrails.custom_rules)) {
-    for (var cr of guardrails.custom_rules) {
-      if (cr.key && cr.description) {
-        rules[cr.key] = { severity: cr.severity || 'medium', description: cr.description };
-      }
-    }
-  }
-
-  updates.rules = rules;
-
-  if (existing) {
-    updateNodeProfile(profileId, updates);
-  } else {
-    createNodeProfile(profileId, Object.assign({ node_type: 'agent', layer: 'customer' }, updates));
-  }
-}
-
-// =============== TEAMS ===============
-
-export function createTeam(id, orgId, name, description, createdBy) {
-  db.prepare(
-    'INSERT INTO teams (id, org_id, name, description, created_by) VALUES (?, ?, ?, ?, ?)'
-  ).run(id, orgId, name, description || '', createdBy || '');
-  return getTeam(id);
-}
-
-export function getTeam(id) {
-  var team = db.prepare('SELECT * FROM teams WHERE id = ?').get(id);
-  if (team) {
-    team.members = db.prepare(
-      'SELECT * FROM team_members WHERE team_id = ? ORDER BY role, joined_at'
-    ).all(id);
-  }
-  return team;
-}
-
-export function listTeams(orgId) {
-  var sql = orgId
-    ? 'SELECT t.*, (SELECT COUNT(*) FROM team_members WHERE team_id = t.id) as member_count FROM teams t WHERE t.org_id = ? ORDER BY t.name'
-    : 'SELECT t.*, (SELECT COUNT(*) FROM team_members WHERE team_id = t.id) as member_count FROM teams t ORDER BY t.name';
-  return orgId ? db.prepare(sql).all(orgId) : db.prepare(sql).all();
-}
-
-export function updateTeam(id, fields) {
-  var changed = buildUpdate('teams', id, fields, ['name', 'description', 'org_id'], { updatedAt: true });
-  if (!changed) return getTeam(id);
-  return getTeam(id);
-}
-
-export function deleteTeam(id) {
-  var memberCount = db.prepare('SELECT COUNT(*) as c FROM team_members WHERE team_id = ?').get(id).c;
-  if (memberCount > 0) throw new Error('Team has members — remove them first');
-  db.prepare('DELETE FROM teams WHERE id = ?').run(id);
-}
-
-export function addTeamMember(teamId, userId, userType, role, isPrimary) {
-  if (isPrimary) {
-    db.prepare('UPDATE team_members SET is_primary = 0 WHERE user_id = ? AND is_primary = 1').run(userId);
-  }
-  db.prepare(
-    'INSERT INTO team_members (team_id, user_id, user_type, role, is_primary) VALUES (?, ?, ?, ?, ?)'
-  ).run(teamId, userId, userType || 'operator', role || 'member', isPrimary ? 1 : 0);
-
-  if (isPrimary) {
-    var table = userType === 'agent' ? 'agents' : 'operators';
-    db.prepare('UPDATE ' + table + ' SET primary_team_id = ? WHERE id = ?').run(teamId, userId);
-  }
-  return db.prepare('SELECT * FROM team_members WHERE team_id = ? AND user_id = ?').get(teamId, userId);
-}
-
-export function updateTeamMember(teamId, userId, fields) {
-  var sets = [];
-  var values = [];
-  if (fields.role) { sets.push('role = ?'); values.push(fields.role); }
-  if (fields.is_primary !== undefined) {
-    if (fields.is_primary) {
-      db.prepare('UPDATE team_members SET is_primary = 0 WHERE user_id = ? AND is_primary = 1').run(userId);
-    }
-    sets.push('is_primary = ?');
-    values.push(fields.is_primary ? 1 : 0);
-  }
-  if (sets.length === 0) return;
-  values.push(teamId, userId);
-  db.prepare('UPDATE team_members SET ' + sets.join(', ') + ' WHERE team_id = ? AND user_id = ?').run(...values);
-
-  if (fields.is_primary) {
-    var member = db.prepare('SELECT user_type FROM team_members WHERE team_id = ? AND user_id = ?').get(teamId, userId);
-    if (member) {
-      var table = member.user_type === 'agent' ? 'agents' : 'operators';
-      db.prepare('UPDATE ' + table + ' SET primary_team_id = ? WHERE id = ?').run(teamId, userId);
-    }
-  }
-}
-
-export function removeTeamMember(teamId, userId) {
-  var member = db.prepare('SELECT * FROM team_members WHERE team_id = ? AND user_id = ?').get(teamId, userId);
-  if (!member) return;
-  db.prepare('DELETE FROM team_members WHERE team_id = ? AND user_id = ?').run(teamId, userId);
-
-  if (member.is_primary) {
-    var table = member.user_type === 'agent' ? 'agents' : 'operators';
-    db.prepare('UPDATE ' + table + ' SET primary_team_id = NULL WHERE id = ?').run(userId);
-  }
-}
-
-export function getTeamsForUser(userId) {
-  return db.prepare(
-    'SELECT t.*, tm.role, tm.is_primary FROM teams t JOIN team_members tm ON t.id = tm.team_id WHERE tm.user_id = ? ORDER BY tm.is_primary DESC, t.name'
-  ).all(userId);
-}
-
-export function getTeamProjects(teamId) {
-  return db.prepare('SELECT * FROM projects WHERE team_id = ?').all(teamId);
-}
-
-
